Use Edge Position property to compute lastIndexOf
Avoid unnecessary loading of the entire collection from the source object.
The incoming edge list is computed from the EReference and iterated to find the
higher Position attribute.

Note: This method as the same behavior as indexOf if the EReference is set
unique (there can be only one incoming edge).