add support for Dynamic EMF (reading)
It is know possible to load / read a model through the Dynamic API
The modification is a simple comparison: if the current factory is
a strict instance of EFactory, then we are using Dynamic EMF, and
EObject reification should be done using PersistenceFactory instead of
standard EFactory.

Dynamic creation of Object can be done with PersistenceFactory.create(EClass),
but it is not possible to use standard EMF methods such as
ECoreUtil.create(EClass), which creates an EObject instance instead of a
PersistentEObject.