Fix subtree containment and non-containment reference issues
Subtree containment issue is related to AddContainmentSubtreeTest test,
and is an inconsistency in the EStore hierarchy that is created if an
entire subtree is added to a resource using its top-level element (only the
top level-elements are added to the resource)

Non-containment reference issue is related to
SetContainmentAfterNonContainmentTest test, and is an EStore / Resource
inconsistency that attach an element that is not contained in the
PersistentResource (using a non-containment reference) to it. This results in
major inconsistency in eContainer computation, and makes element unreachable,
even if they have been added using a containment reference afterwards. This
commit fix #19, and validates it is only related to the Graph implementation.