add a static dependency to EMF Compare
This dependency is used to compile the plugin, but for now it is not used at
runtime because EMF Compare is only available in an Eclipse environment, and
thus is marked as an optional dependency in core plugin. The framework should
use the EMF Compare version available in the Eclipse environment.

Note: EMF Compare has a dependency to Guava [11,16), but NeoEMF uses Guava 21,
this can create some classloading issues and NoSuchMethodError that will be
fixed among time.