use external double linkedlist to represent collections
This implementation allows to compute add, get, and remove operations in O(n/2),
overcoming the limitations of massive remove for large collections.
Anyway some work still have to be done, because having external list creates a
large number of nodes (that are indexed for now), dropping the performances of
vertex access.

Integrating the linked list as a collection of edges between the existing
vertices would help on that.