Fix StackOverflowError when calling toString() on generated EClass metaclass
This error occurs only when NeoEMF code is generated from the Ecore metamodel.
The EClass.toString() method calls the super implementation defined in
DefaultPersistentEObject, which computes the String to return by accessing the
EClass of the current element. This was an issue for ecore models stored in
NeoEMF. Instances of Ecore metamodels are not concerned.